[ICU-3222] Fix some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/14338 Former-commit-id: 67a1c1bd11a43b9d7d869a3a864f84c47ebcf891

+++ icu4c/source/test/intltest/apicoll.cpp
@@ -1216,8 +1216,7 @@ void CollationAPITest::TestMaxExpansion()
         if (order == 0)
             order = iter->previous(status);
 
-        while (U_SUCCESS(status) &&
-            ((uint32_t)iter->previous(status) != UCOL_NULLORDER)) {
+        while (U_SUCCESS(status) && iter->previous(status) != UCOL_NULLORDER) {
             count ++;
         }
